Reorganize String extensions
