+++ iosHyperskillApp/iosHyperskillApp/Sources/Modules/Profile/ViewData/ProfileViewDataMapper.swift
@@ -12,7 +12,7 @@ final class ProfileViewDataMapper {
         let role = profile.isStaff ? Strings.Profile.roleStaff : Strings.Profile.roleLearner
 
         let livesInText: String? = {
-            guard let countryCode = trimmedNonEmptyString(profile.country) else {
+            guard let countryCode = profile.country?.trimmedNonEmptyOrNil() else {
                 return nil
             }
 
@@ -41,19 +41,19 @@ final class ProfileViewDataMapper {
         let socialAccounts: [ProfileSocialAccount] = {
             var result = [ProfileSocialAccount]()
 
-            if let facebookUsername = trimmedNonEmptyString(profile.facebookUsername) {
+            if let facebookUsername = profile.facebookUsername.trimmedNonEmptyOrNil() {
                 result.append(.init(type: .facebook, username: facebookUsername))
             }
-            if let twitterUsername = trimmedNonEmptyString(profile.twitterUsername) {
+            if let twitterUsername = profile.twitterUsername.trimmedNonEmptyOrNil() {
                 result.append(.init(type: .twitter, username: twitterUsername))
             }
-            if let linkedinUsername = trimmedNonEmptyString(profile.linkedinUsername) {
+            if let linkedinUsername = profile.linkedinUsername.trimmedNonEmptyOrNil() {
                 result.append(.init(type: .linkedIn, username: linkedinUsername))
             }
-            if let redditUsername = trimmedNonEmptyString(profile.redditUsername) {
+            if let redditUsername = profile.redditUsername.trimmedNonEmptyOrNil() {
                 result.append(.init(type: .reddit, username: redditUsername))
             }
-            if let githubUsername = trimmedNonEmptyString(profile.githubUsername) {
+            if let githubUsername = profile.githubUsername.trimmedNonEmptyOrNil() {
                 result.append(.init(type: .github, username: githubUsername))
             }
 
@@ -61,24 +61,14 @@ final class ProfileViewDataMapper {
         }()
 
         return ProfileViewData(
-            avatarSource: trimmedNonEmptyString(profile.avatar),
+            avatarSource: profile.avatar.trimmedNonEmptyOrNil(),
             fullname: profile.fullname,
             role: role,
             livesInText: livesInText,
             speaksText: speaksText,
-            bio: trimmedNonEmptyString(profile.bio),
-            experience: trimmedNonEmptyString(profile.experience),
+            bio: profile.bio.trimmedNonEmptyOrNil(),
+            experience: profile.experience.trimmedNonEmptyOrNil(),
             socialAccounts: socialAccounts
         )
-    }
-
-    private func trimmedNonEmptyString(_ stringOrNil: String?) -> String? {
-        guard let string = stringOrNil else {
-            return nil
-        }
-
-        let trimmedString = string.trimmed()
-
-        return trimmedString.isEmpty ? nil : trimmedString
     }
 }
